refactor: extract plan and utils modules to keep lib.rs under 1000 lines
Move plan types (RateSource, CalculationPlan) and logic to src/plan.rs.
Move helper functions (generate_issue_link, truncate, urlencoding_encode)
to src/utils.rs. lib.rs is now 978 lines, well under the 1000-line CI limit.
